fix_database
