+++ BE_Websocket/src/main/java/com/be_websocket/controller/ChatController.java
@@ -1,7 +1,7 @@
 package com.be_websocket.controller;
 
 
-import com.be_websocket.model.Message;
+import com.be_websocket.model.Chat;
 
 import com.be_websocket.service.IUserService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,8 +15,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.Objects;
-
 @RestController
 @RequestMapping("api/chat")
 public class ChatController {
@@ -26,7 +24,7 @@ public class ChatController {
     @Autowired
     IUserService userService;
     @MessageMapping("/sendMessage")
-    public ResponseEntity<?> sendMessage(@Payload Message message) {
+    public ResponseEntity<?> sendMessage(@Payload Chat message) {
         return ResponseEntity.ok(message);
     }
 
@@ -35,10 +33,4 @@ public class ChatController {
         return ResponseEntity.ok(userService.findAll());
     }
 
-//    @MessageMapping("/private-message")
-//    public Message recMessage(@Payload Message message){
-//        simpMessagingTemplate.convertAndSendToUser(message.getReceiver(),"/private",message);
-//        System.out.println(message.toString());
-//        return message;
-//    }
 }
