+++ BE_Websocket/src/main/java/com/be_websocket/model/Room.java
@@ -7,23 +7,26 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
-import java.util.Date;
+import java.util.List;
 
 @Entity
-@Table(name = "messages")
+@Table(name = "rooms")
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class Message {
+public class Room {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+    @ManyToMany
+    @JoinTable(name="user_room",
+    joinColumns = @JoinColumn(name = "room_id"),
+    inverseJoinColumns = @JoinColumn(name = "user_id"))
+    private List<User> userList;
 
-    @ManyToOne
-    @JsonIgnoreProperties({"messages"})
-    private User sender;
-    private String content;
-    private Date timestamp;
-    private Status status;
+    @OneToMany(fetch = FetchType.EAGER, mappedBy = "room")
+    @JsonIgnoreProperties({"room"})
+    private List<Chat> chatList;
+
 }
