+++ BE_Websocket/src/main/java/com/be_websocket/repository/ChatRoomRepository.java
@@ -1,0 +1,9 @@
+package com.be_websocket.repository;
+
+import com.be_websocket.model.Room;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ChatRoomRepository extends JpaRepository<Room, Long> {
+}
