+++ BE_Websocket/src/main/java/com/be_websocket/service/IChatRoomService.java
@@ -1,0 +1,7 @@
+package com.be_websocket.service;
+
+import com.be_websocket.model.Room;
+
+public interface IChatRoomService {
+    Room save(Room room);
+}
