+++ FE_websocket/fe_websocket/src/API/services/chatRoomService.jsx
@@ -1,0 +1,6 @@
+import instance from "../axios"
+
+export const createChatRoom = async (data) => {
+    let response = await instance.post('/api/chat/createChatRoom', data);
+    return response.data;
+}
