+++ FE_websocket/fe_websocket/src/assets/css/login_register.css
@@ -13,7 +13,7 @@ body {
 	flex-direction: column;
 	font-family: 'Montserrat', sans-serif;
 	height: 100vh;
-	margin: -20px 0 50px;
+	/* margin: -20px 0 50px; */
 }
 
 h1 {
