+++ FE_websocket/fe_websocket/src/assets/css/messages.css
@@ -95,7 +95,7 @@ body {
   padding: 30px 30px 20px;
   border-bottom: 2px solid white;
   overflow-y: scroll;
-  height: 500px;
+  height: 300px;
 }
 .chat .chat-history .message-data {
   margin-bottom: 15px;
