+++ FE_websocket/fe_websocket/src/redux/reducers/chatRoomSlice.jsx
@@ -1,0 +1,18 @@
+import { createSlice } from "@reduxjs/toolkit"
+const chatRoomSlice = createSlice(
+    {
+        name: 'chatRoom',
+        initialState: {
+            listChatRoom: [],
+            selectChatRoom: null
+        },
+        reducers: {
+            createChatRoom: (state, action) => {
+
+            }
+        }
+    }
+)
+
+export default chatRoomSlice.reducer
+export const { createChatRoom } = chatRoomSlice.actions
